chore: add missing tests
