+++ test/integrations/destinations/fb_custom_audience/router/record.ts
@@ -1,7 +1,202 @@
-import { Destination, RouterTransformationRequest } from '../../../../../src/types';
+import { Connection, Destination, RouterTransformationRequest } from '../../../../../src/types';
+import { VDM_V2_SCHEMA_VERSION } from '../../../../../src/v0/util/constant';
 import { generateMetadata } from '../../../testUtils';
 
-export const destination: Destination = {
+const destinationV2: Destination = {
+  Config: {
+    accessToken: 'ABC',
+    disableFormat: false,
+    isHashRequired: true,
+    isRaw: false,
+    oneTrustCookieCategories: [],
+    skipVerify: false,
+    subType: 'NA',
+    type: 'NA',
+  },
+  ID: '1mMy5cqbtfuaKZv1IhVQKnBdVwe',
+  Name: 'FB_CUSTOM_AUDIENCE',
+  Enabled: true,
+  WorkspaceID: '1TSN08muJTZwH8iCDmnnRt1pmLd',
+  DestinationDefinition: {
+    ID: '1aIXqM806xAVm92nx07YwKbRrO9',
+    Name: 'FB_CUSTOM_AUDIENCE',
+    DisplayName: 'FB_CUSTOM_AUDIENCE',
+    Config: {},
+  },
+  Transformations: [],
+  IsConnectionEnabled: true,
+  IsProcessorEnabled: true,
+};
+
+const connection: Connection = {
+  sourceId: '2MUWghI7u85n91dd1qzGyswpZan',
+  destinationId: '1mMy5cqbtfuaKZv1IhVQKnBdVwe',
+  enabled: true,
+  config: {
+    destination: {
+      schemaVersion: VDM_V2_SCHEMA_VERSION,
+      disableFormat: false,
+      isHashRequired: true,
+      isRaw: false,
+      subType: 'NA',
+      type: 'NA',
+      audienceId: '23848494844100489',
+    },
+  },
+};
+
+const missingAudienceIDConnection: Connection = {
+  sourceId: '2MUWghI7u85n91dd1qzGyswpZan',
+  destinationId: '1mMy5cqbtfuaKZv1IhVQKnBdVwe',
+  enabled: true,
+  config: {
+    destination: {
+      schemaVersion: VDM_V2_SCHEMA_VERSION,
+    },
+  },
+};
+
+const invalidV2Connection: Connection = {
+  sourceId: '2MUWghI7u85n91dd1qzGyswpZan',
+  destinationId: '1mMy5cqbtfuaKZv1IhVQKnBdVwe',
+  enabled: true,
+  config: {},
+};
+
+export const rETLRecordV2RouterRequest: RouterTransformationRequest = {
+  input: [
+    {
+      destination: destinationV2,
+      connection: connection,
+      message: {
+        action: 'insert',
+        context: {
+          sources: {
+            job_run_id: 'cgiiurt8um7k7n5dq480',
+            task_run_id: 'cgiiurt8um7k7n5dq48g',
+            job_id: '2MUWghI7u85n91dd1qzGyswpZan',
+            version: '895/merge',
+          },
+        },
+        recordId: '2',
+        rudderId: '2',
+        identifiers: {
+          EMAIL: 'subscribed@eewrfrd.com',
+          FI: 'ghui',
+        },
+        type: 'record',
+      },
+      metadata: generateMetadata(1),
+    },
+    {
+      destination: destinationV2,
+      connection: connection,
+      message: {
+        action: 'insert',
+        context: {
+          sources: {
+            job_run_id: 'cgiiurt8um7k7n5dq480',
+            task_run_id: 'cgiiurt8um7k7n5dq48g',
+            job_id: '2MUWghI7u85n91dd1qzGyswpZan',
+            version: '895/merge',
+          },
+        },
+        recordId: '2',
+        rudderId: '2',
+        identifiers: {
+          EMAIL: 'subscribed@eewrfrd.com',
+          FI: 'ghui',
+        },
+        type: 'record',
+      },
+      metadata: generateMetadata(2),
+    },
+    {
+      destination: destinationV2,
+      connection: connection,
+      message: {
+        action: 'insert',
+        context: {
+          sources: {
+            job_run_id: 'cgiiurt8um7k7n5dq480',
+            task_run_id: 'cgiiurt8um7k7n5dq48g',
+            job_id: '2MUWghI7u85n91dd1qzGyswpZan',
+            version: '895/merge',
+          },
+        },
+        recordId: '2',
+        rudderId: '2',
+        identifiers: {
+          EMAIL: 'subscribed@eewrfrd.com',
+          FI: 'ghui',
+        },
+        type: 'record',
+      },
+      metadata: generateMetadata(3),
+    },
+  ],
+  destType: 'fb_custom_audience',
+};
+
+export const rETLRecordV2RouterInvalidRequest: RouterTransformationRequest = {
+  input: [
+    {
+      destination: destinationV2,
+      connection: missingAudienceIDConnection,
+      message: {
+        action: 'insert',
+        context: {
+          sources: {
+            job_run_id: 'cgiiurt8um7k7n5dq480',
+            task_run_id: 'cgiiurt8um7k7n5dq48g',
+            job_id: '2MUWghI7u85n91dd1qzGyswpZan',
+            version: '895/merge',
+          },
+        },
+        recordId: '2',
+        rudderId: '2',
+        identifiers: {
+          EMAIL: 'subscribed@eewrfrd.com',
+          FI: 'ghui',
+        },
+        type: 'record',
+      },
+      metadata: generateMetadata(1),
+    },
+  ],
+  destType: 'fb_custom_audience',
+};
+
+export const rETLRecordInvalidRequest: RouterTransformationRequest = {
+  input: [
+    {
+      destination: destinationV2,
+      connection: invalidV2Connection,
+      message: {
+        action: 'insert',
+        context: {
+          sources: {
+            job_run_id: 'cgiiurt8um7k7n5dq480',
+            task_run_id: 'cgiiurt8um7k7n5dq48g',
+            job_id: '2MUWghI7u85n91dd1qzGyswpZan',
+            version: '895/merge',
+          },
+        },
+        recordId: '2',
+        rudderId: '2',
+        identifiers: {
+          EMAIL: 'subscribed@eewrfrd.com',
+          FI: 'ghui',
+        },
+        type: 'record',
+      },
+      metadata: generateMetadata(1),
+    },
+  ],
+  destType: 'fb_custom_audience',
+};
+
+export const destinationV1: Destination = {
   Config: {
     accessToken: 'ABC',
     appSecret: 'dummySecret',
@@ -33,7 +228,7 @@ export const destination: Destination = {
 export const rETLRecordV1RouterRequest: RouterTransformationRequest = {
   input: [
     {
-      destination: destination,
+      destination: destinationV1,
       message: {
         action: 'insert',
         context: {
@@ -63,7 +258,7 @@ export const rETLRecordV1RouterRequest: RouterTransformationRequest = {
       metadata: generateMetadata(3),
     },
     {
-      destination: destination,
+      destination: destinationV1,
       message: {
         action: 'update',
         context: {
@@ -93,7 +288,7 @@ export const rETLRecordV1RouterRequest: RouterTransformationRequest = {
       metadata: generateMetadata(4),
     },
     {
-      destination: destination,
+      destination: destinationV1,
       message: {
         action: 'delete',
         context: {
@@ -123,7 +318,7 @@ export const rETLRecordV1RouterRequest: RouterTransformationRequest = {
       metadata: generateMetadata(1),
     },
     {
-      destination: destination,
+      destination: destinationV1,
       message: {
         action: 'delete',
         context: {
@@ -153,7 +348,7 @@ export const rETLRecordV1RouterRequest: RouterTransformationRequest = {
       metadata: generateMetadata(2),
     },
     {
-      destination: destination,
+      destination: destinationV1,
       message: {
         action: 'update',
         context: {
@@ -183,7 +378,7 @@ export const rETLRecordV1RouterRequest: RouterTransformationRequest = {
       metadata: generateMetadata(5),
     },
     {
-      destination: destination,
+      destination: destinationV1,
       message: {
         action: 'update',
         context: {
@@ -213,7 +408,7 @@ export const rETLRecordV1RouterRequest: RouterTransformationRequest = {
       metadata: generateMetadata(6),
     },
     {
-      destination: destination,
+      destination: destinationV1,
       message: {
         action: 'lol',
         context: {
@@ -244,8 +439,4 @@ export const rETLRecordV1RouterRequest: RouterTransformationRequest = {
     },
   ],
   destType: 'fb_custom_audience',
-};
-
-module.exports = {
-  rETLRecordV1RouterRequest,
 };
